re-hierarch definitions
