+++ squeal-postgresql/src/Squeal/PostgreSQL/Migration.hs
@@ -167,6 +167,7 @@ import qualified GHC.Generics as GHC
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Binary
 import Squeal.PostgreSQL.Definition
+import Squeal.PostgreSQL.Definition.Table
 import Squeal.PostgreSQL.Expression.Comparison
 import Squeal.PostgreSQL.Expression.Parameter
 import Squeal.PostgreSQL.Expression.Time
